Reordered appointment service
The base of appointment service is disordered. Now, it has been fixed:
- Directories have been restructured
- Removed unused code that causes error
- Fixed pom file details
- Added banner, application info, and entry point
