+++ appointment/src/main/java/com/github/coding_team_sept/nd_backend/appointment/AppointmentApplication.java
@@ -1,0 +1,11 @@
+package com.github.coding_team_sept.nd_backend.appointment;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AppointmentApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(AppointmentApplication.class, args);
+    }
+}
