+++ appointment/src/main/java/com/github/coding_team_sept/nd_backend/appointment/payload/requests/AppointmentRequest.java
@@ -1,0 +1,5 @@
+package com.github.coding_team_sept.nd_backend.appointment.payload.requests;
+
+public class AppointmentRequest {
+
+}
