+++ appointment/src/main/java/com/github/coding_team_sept/nd_backend/appointment/repository/AppointmentRepository.java
@@ -1,0 +1,9 @@
+//package com.github.coding_team_sept.nd_backend.appointment.repository;
+//
+//import org.springframework.data.jpa.repository.JpaRepository;
+//import org.springframework.stereotype.Repository;
+//
+//@Repository
+//public interface AppointmentRepository extends JpaRepository<Object, Long> {
+//    // TODO: Create repository for appointment
+//}
