+++ appointment/src/main/java/com/github/coding_team_sept/nd_backend/appointment/service/AppointmentService.java
@@ -1,0 +1,8 @@
+package com.github.coding_team_sept.nd_backend.appointment.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class AppointmentService {
+    // TODO: Create appointment service for appointment
+}
