--- appointment/src/main/java/com/github/coding_team_sept/nd_backend/model/DoctorUser.java
@@ -1,26 +1,0 @@
-package com.github.coding_team_sept.nd_backend.model;
-
-
-import org.springframework.data.annotation.Id;
-
-import javax.persistence.Column;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.SequenceGenerator;
-
-public class DoctorUser {
-    @Id
-    @SequenceGenerator(
-            name = "user_sequence_id",
-            sequenceName = "user_sequence_id"
-    )
-    @GeneratedValue(
-            strategy = GenerationType.SEQUENCE,
-            generator = "user_sequence_id"
-    )
-    private Long id;
-    @Column(nullable = false, unique = true)
-    private String name;
-    @Column(nullable = false)
-    private String time;
-    }
