--- appointment/src/main/java/com/github/coding_team_sept/nd_backend/payload/requests/AppointmentRequest.java
@@ -1,5 +1,0 @@
-package com.github.coding_team_sept.nd_backend.payload.requests;
-
-public class AppointmentRequest {
-
-}
