+++ appointment/src/main/resources/application.yml
@@ -1,0 +1,21 @@
+server:
+  port: 9001
+
+spring:
+  application:
+    name: appointment
+
+  datasource:
+    username: 'admin'
+    url: jdbc:postgresql://localhost:5432/telemedic # TODO: Update URL
+    password: 'admin'
+
+  jpa:
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.PostgreSQLDialect
+        format_sql: 'true'
+    hibernate:
+      ddl-auto: update
+    show-sql: 'true'
+
